+++ SurfsUpv3/Views/Shared/_Layout.cshtml
@@ -12,7 +12,7 @@
     <link rel="stylesheet" href="~/css/_navbar.css" />
     <link rel="stylesheet" href="~/css/_CatelogView.css" />
     <link rel="stylesheet" href="~/css/_InsertBookingDetails.css" />
-    <link rel="stylesheet" href="~/css/_AdminViews.css" />
+   @*  <link rel="stylesheet" href="~/css/_AdminViews.css" /> *@
     <!-- <link rel="stylesheet" href="~/css/_Footer.css" /> -->
     @* Grab from external Google API to get the surfy-font, 'Matemasie' *@
     <link href="https://fonts.googleapis.com/css2?family=Matemasie&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" rel="stylesheet" />
