+++ SurfsUpv3/Views/Udlejning/Index.cshtml
@@ -1,5 +1,5 @@
-﻿@using SurfsUpv3.Models
-@model IEnumerable<Surfboard>
+﻿
+@model IEnumerable<SurfsUpv3.Models.Surfboard>
 
 
 
@@ -43,61 +43,3 @@
 
 </div>
 
-
-
-<br />
-<div class="main-container">
-    <div class="product-list">
-        <h3>Våddragter</h3><br />
-        <div class="row">
-            <div class="product-container">
-                <div class="product-wetsuit">
-                    <button class="btn btn-primary wetsuit-item d-block text-center py-2">
-                        <img src="\images\VåddragtBlue.jpg" class="img-fluid" /><br />
-                        Lej en våddragt
-                    </button>
-                    <button class="btn btn-primary wetsuit-item d-block text-center py-2">
-                        <img src="\images\VåddragtBlue.jpg" class="img-fluid" /> <br />
-                        Lej en våddragt
-                    </button>
-                    <button class="btn btn-primary wetsuit-item d-block text-center py-2">
-                        <img src="\images\VåddragtBlue.jpg" class="img-fluid" /><br />
-                        Lej en våddragt
-                    </button>
-                </div>
-            </div>
-
-
-        </div>
-    </div>
-</div>
-<br />
-
-<div class="container">
-    <h3>Andre ting</h3><br />
-    <div class="row">
-        <div class="col-4">
-            <button class="Surfboard">
-                <img src="\images\LifeJacket.jpg" width="150px"
-                     height="150px"><br />
-                Lej redningsvest
-            </button>
-
-        </div>
-        <div class="col-4">
-            <button class="Surfboard">
-                <img src="\images\SurfGloves.jpg" width="100px"
-                     height="100px"><br />
-                Lej handsker
-            </button>
-
-        </div>
-        <div class="col-4">
-            <button class="Surfboard">
-                <img src="\images\PaddleBlack.jpg" width="100px"
-                     height="100px"><br />
-                Lej roudstyr
-            </button>
-        </div>
-    </div>
-</div>
